test: Rename test_definitions to utils
