+++ test/test_cli.py
@@ -8,7 +8,7 @@ import subprocess
 from cve_bin_tool.cli import main
 from cve_bin_tool.extractor import Extractor
 
-from .test_definitions import (
+from .utils import (
     TempDirTest,
     download_file,
     CURL_7_20_0_RPM,
