+++ test/test_extract.py
@@ -8,7 +8,7 @@ import unittest
 from zipfile import ZipFile, ZipInfo
 from io import BytesIO
 from cve_bin_tool.extractor import Extractor
-from .test_definitions import download_file, CURL_7_20_0_URL, VMWARE_CAB, TMUX_DEB
+from .utils import download_file, CURL_7_20_0_URL, VMWARE_CAB, TMUX_DEB
 from cve_bin_tool.util import inpath
 
 if sys.version_info.major == 3 and sys.version_info.minor >= 3:
