+++ test/test_scanner.py
@@ -11,7 +11,7 @@ from sys import platform
 
 from cve_bin_tool.cvedb import CVEDB
 from cve_bin_tool.cli import Scanner, InvalidFileError
-from .test_definitions import download_file
+from .utils import download_file
 
 BINARIES_PATH = os.path.join(os.path.abspath(os.path.dirname(__file__)), "binaries")
