+++ test/utils.py
@@ -1,5 +1,5 @@
 """
-Resource definitions and helper functions for tests.
+Resource definitions and helper utilities for tests.
 """
 import shutil
 import tempfile
